Added session control to modelIndex view.
