+++ lib/router.js
@@ -41,8 +41,8 @@ exports.setupModelRoutes = function () {
 
 	// model routes
 	router.get('/models/list', routes.modelList);
-	router.get('/model/:model/list', middleware.modelIndex(), routes.modelIndex);
-    router.post('/model/:model/list', middleware.modelIndex(), routes.modelIndex);
+	router.get('/model/:model/list', middleware.modelIndex.get, routes.modelIndex);
+    router.post('/model/:model/list', middleware.modelIndex.post, routes.modelIndex);
 	router.get('/model/:model/new', middleware.modelCreate(), routes.modelCreate);
 	router.post('/model/:model/create', middleware.modelSave(), routes.modelSave);
